Finally settled on what looks like a pretty reasonable approach for u_nk dfs
We are now including lambda vector values as index columns in `u_nk`
DataFrames. This means we don't have to care about order of
concatenation generally, since an estimator such as MBAR can handle
sorting on its own. This is a huge usability boon, but it does add some
(perhaps mostly perceived) complexity to the `u_nk` data structure.
We will need to document this clearly, since all parsers will need to
adhere to this standard.