omit _version.py from coverage reporting
The _version.py file is installed by versioneer and we never touch it.
It is difficult to test exhaustively. Basic tests are in test_version.py
but they only get about 36% coverage. We assume that versioneer does the
correct thing and do not include it in the coverage that WE are
responsible for.