Changes in light of upstream pymbar 3.0.5
We need an explicit `return_theta=True` call now, as before this was
implicitly working.

Also switched to using a list for alchemlyb.estimators.MBAR.method, as
this is in line with the stated docs for inputs to pymbar.MBAR (we were
using a tuple).

We have set the install requirements in setupy.py to now have
`pymbar==3.0.5` to ensure users are staying up to the latest version,
as we intend to track pymbar's development carefully.