First iteration of bootstrapping function
Should work on both dHdl and u_nk standard forms.
Still requires:
- tests
- usage example(s)

We additionally want to add a bootstrapping keyword to each estimator's
constructor, which changes how uncertainties are estimated at the
expense of more compute.
This will be the approach we will push users toward using for
simplicity, but we still want to expose building blocks by way of the
`bootstrap` function to allow for scaling out if needed (e.g. with
dask).