Added tests for subsampling on most of our datasets
Not getting all passes yet, and not sure I'm happy with how these work.
There are clear shortcomings with the static column approach, requiring
perhaps manual effort on the part of the user to work around.
This is counter to alchemlyb's philosophy of making the 'most right'
thing easy to do.