Docs for subsamplers reflect a change in direction
In the spirit of wanting to make it easy for users to do the 'most
right' thing, I have looked at the details of what `alchemical-analysis`
currently does for subsampling with statisticalInefficiency, and it uses
an approach similar to the one documented in each of our
`statistical_inefficiency` and `equilbrium_detection` functions now.

Next steps are to *actually* implement this functionality.
We are going to make our lives easier by adding an `alchemform`
attribute to dataframes our parsers produce, and add this to the
standard form spec. This attribute will have 'dHdl' or 'u_nk' as its
value, depending on the form of data we're dealing with.
Our estimators can also use this as a check or warning.