alchemform hits a stumbling block
Using something like `pd.concat` on a set of DataFrames doesn't
propagate forward the `.attrs` dictionary, so having an `alchemform` key
here created by the parsers is not sufficient for signalling to
preprocessors the type of data.

We will likely remove the 'auto' option from the subsamplers and insist
on explicit values for `how`. This is not *necessarily* bad, as it
promotes some understanding for the user of what they are choosing to do
here.