Sort filenames by basename; error on non-restarted truncated window
We sort fepout filenames by basename because they may be in different directories,
especially in the tests where some of them are modified on the fly.

While this parser is meant to handle restarted runs, it doesn't (yet) handle
the case where a window is just truncated and the end of window is implied
by end of file.